fix: harden core/protocol wire compatibility
Lenient RepeatMode decode (unknown/empty -> Off) so a legacy or
MCP-forwarded value can't kill the IPC frame; serialization stays
canonical lowercase.

Lenient MediaItem.release_date field decode: a malformed cached/provider
date (e.g. "0000-00-00") becomes None instead of failing the payload.

Make ReleaseDate fields private with accessors and degrade Display for
the day-without-month case, removing the unreachable! panic Serialize
could route through mid-encode.

Fail-open MusicProvider::playlist_version_changed default: a missing
token on either side reports changed, matching the sync contract.

Replace u128 with u64 in serde/IPC payloads (provider_api_finished_event,
RebuildReport.elapsed_ms) per the no-u128-in-serde rule.

Extract PREMIUM_REQUIRED_POLICY_REASON constant in core, shared by the
protocol downgrade path and the player, so the two can't drift.

Document that non-spotify SearchSourceData::Remote values aren't
decodable by released daemons, and that id-newtype new() is an
unvalidated escape hatch.

+++ crates/spotuify-core/src/analytics.rs
@@ -150,7 +150,7 @@ pub fn provider_api_finished_event(
     method: &str,
     path: &str,
     status: Option<u16>,
-    elapsed_ms: u128,
+    elapsed_ms: u64,
     error_class: Option<&str>,
     occurred_at_ms: i64,
 ) -> AnalyticsEvent {
